22:06 3/17/13
Finished work on the EnumArea Class, excepting a massive shift of my
understanding of the code, that class is now completely finished. New
Areas can now be templated within that class by adding new entries to
the list of enums at the top of the class. Next Props will be built,
likely using Enums as well, and items will be written in, at which point
a test dungeon can be made and we will enter Alpha v0.1.0.