16:09 3/19/13
Wrote a function that describes props in an area called describeProps()
in Area. Uses a hashmap of the props to create a set of keys that it
iterates through and builds a return from the props' enums. First time
working with Map, Set and Iterator, no chance to test until I reach
Alpha 0.1.0 so we'll see if it works then.