20:58 3/20/13
Finished the area, prop and item generation systems for non randomized
objects. Next is either randomization, the player systems or command
systems. I don't know which way I'm leaning at the moment but the
classes for randomization lists are in, so maybe that. Also might start
learning swing and using that as a graphic engine so that the system
isn't quite as crap, that's for the future though, likely after entering
Alpha and getting basic mechanics and major bugs worked out.