20:39 4/2/13
Been a while since I've pushed. I've worked out Exits and moving,
unfortunately it revealed a serious bug in the core of World that causes
Null Pointers for all of the world spaces. This will be fixed before the
next push. Player has also been expanded slightly, and there have been
minor tweaks to Areas, including the addition of a new area type, a
dropping of the limit on exits in an area and the function
addExitToArea.