22:38 5/26/13
Dropped the project for a while, now back into it. Rewrote probably 90%
of the code, still getting back to where I was, but it was much easier
with the old code serving as a guide. Dropped the old Map indexing
convention of the key being the value plus a number. Every single object
except for areas is now given a unique id number generated from the
UniqueID class which is universally used to call it through maps. Areas
now also work in a more intuitive way, with all areas being contained in
submaps of other areas with the overworld/exteriors all being in the
overworld dummy area, saves calculation time and headache while coding.
In addition, areas now move items between maps rather than creating new
instances of the item every time it moves, this will require extra focus
to what functions they are using for the coder, but save memory. Areas
All uniqueId objects work in a very similar fashion now as opposed to
the old way where everything had it's own quirks that made them a bitch
to code. Creatures have been changed, they can no longer reference their
own containing Area for the moment, as it leads to a problem with the
new handling method. Once Time is added all things which change on a
regular basis (creatures with AI, rotting food items, etc.) will poll
time to add themselves to the queue to change when their time comes.
Until such time monsters and other animals will simply wait in their
home Areas to be killed, and move through areas to chase the player.