20:50 5/27/13
Command is back with a very different functionality, now it is simply a
central processor for scripts with ties to every object with a UniqueID,
similar to the initial plan for it back in the early pre-rewrite. Most
objects will use generic scripts based on their types that allow them to
be examined, picked up, etc. based on what is appropriate for their
type. As a part of the new script and command system, all objects with a
UniqueID now extend the class UniqueIDObject which acts as a wrapper
data type so that the Unique objects, as I've taken to calling such
objects, can be referred to generically, while also allowing me to refer
to program specific methods within them such as getScript().